Revert "trying to update and add ember-hook.test-loader issues"
This reverts commit a8504ac531f1d16f27ebe1cc0029cb13e5390216.
