+++ README.md
@@ -13,35 +13,44 @@
 ```
 ember install ember-frost-info-bar
 ```
-## Slots API
-
-Using [ember-block-slots](https://github.com/ciena-blueplanet/ember-block-slots), this component will provide an icon, title and subtitle area, controls area for components like Search and actions slot.
 
 ## Examples
 ### Block Format
 ```handlebars
-{{#frost-info-bar}}
-  {{#block-slot 'icon'}}
-    {{frost-icon
-      icon='bacon'
-      pack='dummy'
-    }}
+{{#frost-info-bar as |slot|}}
+  {{#block-slot slot 'title'}}
+      User accounts
   {{/block-slot}}
-  {{#block-slot 'title'}}
-    &lt;placeholder: title&gt;
+  {{#block-slot slot 'summary'}}
+      Summary section
   {{/block-slot}}
-  {{#if summary}}
-    {{#block-slot 'summary'}}
-      &lt;placeholder: summary&gt;
-    {{/block-slot}}
-  {{/if}}
-  {{#block-slot 'controls'}}
-      &lt;placeholder: controls&gt;
+  {{#block-slot slot 'context-controls'}}
+      Context controls section
   {{/block-slot}}
-  {{#block-slot 'actions' as |action|}}
-    {{action.button icon='frost/infobar-create' text='Click me!' onClick=(action 'triggerAction')}}
+  {{#block-slot slot 'actions' as |action|}}
+    {{action.button icon='frost/infobar-create' text='Create' onClick=(action 'createUser')}}
   {{/block-slot}}
 {{/frost-info-bar}}
+```
+
+### HTML Format
+```handlebars
+<div class='frost-info-bar'>
+    <div class='title'>
+        <div class='primary-title'>User accounts</div>
+        <div class='sub-title'>View and manage user accounts</div>
+    </div>
+    <div class='action'>
+    {{frost-button
+      onClick=(action 'createUser')
+      priority="tertiary"
+      size="medium"
+      icon="frost/infobar-create"
+      text="Create"
+      vertical=true
+    }}
+    </div>
+</div>
 ```
 
 ## Development
