+++ bower.json
@@ -1,22 +1,23 @@
 {
   "name": "ember-frost-info-bar",
   "dependencies": {
-    "ember": "~2.6.2",
-    "ember-cli-shims": "0.1.1",
-    "ember-cli-test-loader": "0.2.2",
+    "ember": "2.4.2",
+    "ember-cli-shims": "ember-cli/ember-cli-shims#0.1.0",
+    "ember-cli-test-loader": "ember-cli/ember-cli-test-loader#0.1.3",
     "jquery": "2.1.4",
     "blanket": "1.1.7",
     "bourbon": "4.2.6",
+    "d3": "3.5.6",
     "ember-inflector": "1.3.1",
+    "ember-mocha": "~0.8.6",
+    "Faker": "3.0.1",
     "lodash": "^4.10.0",
     "perfect-scrollbar": ">=0.6.7 <2.0.0",
-    "sinonjs": "1.17.1",
-    "mocha": "~2.2.4",
-    "chai": "~2.3.0",
-    "ember-mocha-adapter": "~0.3.1"
+    "pretender": "0.10.1",
+    "sinonjs": "1.17.1"
   },
   "resolutions": {
-    "ember": "~2.6.2"
+    "ember": "2.4.2"
   },
   "devDependencies": {
     "blanket": "5e94fc30f2e694bb5c3718ddcbf60d467f4b4d26"
