+++ ember-cli-build.js
@@ -6,9 +6,9 @@ module.exports = function (defaults) {
     'ember-cli-mocha': {
       useLintTree: false
     },
-    babel: {
-      optional: ['es7.decorators'],
-      ignore: ['node_modules/**/*.js', 'bower_components/**/*.js']
+    sassOptions: {
+      includePaths: [
+      ]
     }
   })
