+++ package.json
@@ -18,40 +18,39 @@
     "node": ">= 5.0.0"
   },
   "author": "Roxanne Panthaky (https://github.com/rox163)",
-  "contributors": [
-    "Matthew Dahl (https://github.com/sandersky)",
-    "Quincy Le (https://github.com/quincyle)",
-    "Olufemi Adeojo (https://github.com/oadeojo)",
-    "Gosia Hyndman (https://github.com/vesper2000)"
-  ],
+  "contributors": [],
   "license": "MIT",
   "devDependencies": {
     "broccoli-asset-rev": "^2.4.2",
     "ember-block-slots": "^1.1.0",
     "ember-browserify": "1.1.11",
-    "ember-cli": "^2.6.3",
+    "ember-cli": "^2.4.2",
     "ember-cli-app-version": "1.0.0",
-    "ember-cli-blanket": "0.9.5",
+    "ember-cli-blanket": "0.9.1",
+    "ember-cli-content-security-policy": "0.5.0",
     "ember-cli-dependency-checker": "^1.2.0",
     "ember-cli-htmlbars-inline-precompile": "0.3.1",
+    "ember-cli-ic-ajax": "0.2.4",
     "ember-cli-inject-live-reload": "1.4.0",
-    "ember-cli-mocha": "0.10.4",
+    "ember-cli-mirage": "0.1.13",
+    "ember-cli-mocha": "0.10.1",
     "ember-cli-notifications": "3.2.0",
     "ember-cli-uglify": "1.2.0",
     "ember-computed-decorators": "0.2.2",
+    "ember-data": "^2.4.0",
     "ember-disable-proxy-controllers": "1.0.1",
     "ember-export-application-global": "1.0.5",
     "ember-frost-core": ">=0.20.0 <1.0.0",
     "ember-frost-notifier": "^2.0.0",
-    "ember-hook": "1.2.1",
     "ember-load-initializers": "^0.5.1",
     "ember-lodash": "0.0.10",
     "ember-one-way-controls": "0.8.3",
     "ember-resolver": "^2.0.3",
     "ember-truth-helpers": "1.2.0",
+    "ember-try": "^0.2.0",
     "eslint": "^2.9.0",
     "eslint-config-frost-standard": "^2.0.0",
-    "loader.js": "^4.0.10",
+    "loader.js": "^4.0.1",
     "svg4everybody": "2.1.0"
   },
   "keywords": [
@@ -59,11 +58,11 @@
     "frost"
   ],
   "dependencies": {
-    "ember-cli-babel": "^5.1.6",
-    "ember-cli-htmlbars": "^1.0.3",
-    "ember-cli-sass": "5.3.1",
+    "ember-cli-babel": "^5.1.5",
+    "ember-cli-htmlbars": "^1.0.1",
+    "ember-cli-sass": "^5.1.0",
     "ember-computed-decorators": "^0.2.2",
-    "liquid-fire": "~0.24.0"
+    "liquid-fire": "^0.23.0"
   },
   "ember-addon": {
     "configPath": "tests/dummy/config"
