+++ tests/dummy/app/app.js
@@ -1,6 +1,6 @@
 import Ember from 'ember'
-import Resolver from './resolver'
-import loadInitializers from 'ember-load-initializers'
+import Resolver from 'ember/resolver'
+import loadInitializers from 'ember/load-initializers'
 import config from './config/environment'
 
 let App
