+++ tests/dummy/app/router.js
@@ -1,7 +1,7 @@
 import Ember from 'ember'
 import config from './config/environment'
 
-const Router = Ember.Router.extend({
+var Router = Ember.Router.extend({
   location: config.locationType
 })
